refactor: ShoppingItem 모델로 전환
