+++ app/src/main/java/woowacourse/shopping/view/products/adapter/ProductAdapter.kt
@@ -9,7 +9,7 @@ import woowacourse.shopping.databinding.ItemProductSkeletonBinding
 import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.view.cartcounter.OnClickCartItemCounter
 import woowacourse.shopping.view.products.OnClickProducts
-import woowacourse.shopping.view.products.ShoppingItem
+import woowacourse.shopping.view.products.model.ShoppingItem
 import woowacourse.shopping.view.products.adapter.viewholder.ProductViewHolder
 import woowacourse.shopping.view.products.adapter.viewholder.SkeletonViewHolder
